matchd: Add Python shell and IES-API wrapper module
This patch adds an embedded Python shell within matchd. It is intended
for debugging and extending matchd and matchdlib. The initial goal is
to provide easy access to IES-API calls, interactively from within
matchd. IES-API provided many debug dump functions that could provide
important information about the underlying IES HW pipeline and useful
statistics.
The Python shell itself is extremely simple (few calls to Python
library). It is provided by libonp - a shared library with generic
functionality such as logging. It could be shared with multiple
applications.

Python shell feature and the IES-API wrapper generation are controlled
by --enable-matchd-pysh configure option.
This feature requires liponp to be installed with --prefix=/usr or a
full path to libonp headers and libraries could be provided with
  --with-libonp-headers=FULLPATH
                      Path to libonp headers (default: prefix/include)
  --with-libonp-libraries=FULLPATH
                      Path to libonp libraries (default: prefix/libs)
When --enable-matchd-pysh is entered, the include path of IES-API
headers must be provided with:
  --with-ies-api-headers=FULLPATH
                      Path to IES API headers (default: prefix/include)
When both libonp and IES-API are installed with --prefix=/usr, matchd
has to be configured using:
./configure --prefix=/usr --sysconfdir=/etc --enable-matchd-pysh \
            --with-ies-api-headers=/usr/include
After building matchd, it should be copied along with the generetaed
fm module files to the working directory of matchd. The following files
have to be copied:
[match-interface-build-top]$ cp src/.libs/matchd src/.libs/_fm.so src/fm.py \
    $(MATCHD_WORKING_DIRECTORY)

matchd has to be started manually (i.e. not as a daemon). It will
initialize the switch and drop into the Python shell prompt ">>>".
matchd's main thread will continue to run and listen to requests from
match tool.

To gain access to IES-API calls from within the Python shell enter:
>>> import os
>>> import sys
>>> sys.path.append(os.getcwd())
>>> import fm
>>> from fm import *

Signed-off-by: Ronen Arad <ronen.arad@intel.com>

Conflicts:
	src/Makefile.am
	src/matchd.c